+++ src/o365/onedrive/commands/report/report-usageaccountcounts.spec.ts
@@ -3,7 +3,7 @@ import Command from '../../../../Command';
 import * as sinon from 'sinon';
 import appInsights from '../../../../appInsights';
 import auth from '../../../../Auth';
-const command: Command = require('./onedrive-report-usageaccountcounts');
+const command: Command = require('./report-usageaccountcounts');
 import * as assert from 'assert';
 import Utils from '../../../../Utils';
 import request from '../../../../request';
